# sila-me/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: SILA Cloud shared backend API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /materials/search:
    get:
      operationId: searchMaterials
      tags: [master-data]
      summary: Search authorized active materials
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: barcode, in: query, schema: { type: string } }
        - { name: plantCode, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 0, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0 } }
      responses:
        "200":
          description: Material search results
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaterialSearchResponse" }
        "403":
          description: Insufficient permission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /materials/by-code/{code}:
    get:
      operationId: getMaterialByCode
      tags: [master-data]
      summary: Resolve an authorized material code
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Material detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaterialDetail" }
        "404":
          description: Material not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Code is ambiguous without a plant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /materials/recognize:
    get:
      operationId: recognizeMaterial
      tags: [master-data]
      summary: Deterministically recognize a material from supplied clues
      parameters:
        - { name: barcode, in: query, schema: { type: string } }
        - { name: materialCode, in: query, schema: { type: string } }
        - { name: supplierMaterialCode, in: query, schema: { type: string } }
        - { name: description, in: query, schema: { type: string } }
        - { name: alias, in: query, schema: { type: string } }
        - { name: brand, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
      responses:
        "200":
          description: Deterministic recognition result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaterialRecognitionResponse" }
  /materials/confirm:
    post:
      operationId: confirmMaterial
      tags: [master-data]
      summary: Confirm a material match and record evidence
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MaterialConfirmationInput" }
      responses:
        "201":
          description: Confirmed material
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaterialDetail" }
  /materials/{id}:
    get:
      operationId: getMaterial
      tags: [master-data]
      summary: Get an authorized material detail
      parameters:
        - { $ref: "#/components/parameters/IntegerId" }
      responses:
        "200":
          description: Material detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaterialDetail" }
        "404":
          description: Material not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/materials:
    post:
      operationId: createMaterial
      tags: [administration, master-data]
      summary: Create a development or provider material
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MaterialInput" }
      responses:
        "201":
          description: Created material
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaterialDetail" }
        "400":
          description: Invalid material
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Insufficient permission or customer scope
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/materials/{id}:
    put:
      operationId: updateMaterial
      tags: [administration, master-data]
      summary: Update an authorized material
      parameters:
        - { $ref: "#/components/parameters/IntegerId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MaterialUpdate" }
      responses:
        "200":
          description: Updated material
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaterialDetail" }
  /admin/materials/{materialId}/aliases:
    post:
      operationId: createMaterialAlias
      tags: [administration, master-data]
      summary: Add a searchable material alias
      parameters:
        - { name: materialId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MaterialAliasInput" }
      responses:
        "201":
          description: Material detail after alias creation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaterialDetail" }
  /admin/materials/{materialId}/aliases/{aliasId}:
    delete:
      operationId: deleteMaterialAlias
      tags: [administration, master-data]
      summary: Remove a material alias
      parameters:
        - { name: materialId, in: path, required: true, schema: { type: integer } }
        - { name: aliasId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Material detail after alias deletion
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaterialDetail" }
  /admin/materials/{materialId}/barcodes:
    post:
      operationId: createMaterialBarcode
      tags: [administration, master-data]
      summary: Add a material barcode
      parameters:
        - { name: materialId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MaterialBarcodeInput" }
      responses:
        "201":
          description: Material detail after barcode creation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaterialDetail" }
  /admin/materials/{materialId}/barcodes/{barcodeId}:
    delete:
      operationId: deleteMaterialBarcode
      tags: [administration, master-data]
      summary: Remove a material barcode
      parameters:
        - { name: materialId, in: path, required: true, schema: { type: integer } }
        - { name: barcodeId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Material detail after barcode deletion
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaterialDetail" }
  /admin/materials/{materialId}/uom-conversions:
    post:
      operationId: upsertMaterialUomConversion
      tags: [administration, master-data]
      summary: Create or update a material UOM conversion
      parameters:
        - { name: materialId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MaterialUomConversionInput" }
      responses:
        "201":
          description: Material detail after UOM conversion update
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaterialDetail" }
  /admin/materials/{materialId}/uom-conversions/{conversionId}:
    delete:
      operationId: deleteMaterialUomConversion
      tags: [administration, master-data]
      summary: Remove a material UOM conversion
      parameters:
        - { name: materialId, in: path, required: true, schema: { type: integer } }
        - { name: conversionId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Material detail after UOM conversion deletion
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaterialDetail" }
  /suppliers/search:
    get:
      operationId: searchSuppliers
      tags: [master-data]
      summary: Search authorized active suppliers
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 0, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0 } }
      responses:
        "200":
          description: Supplier search results
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupplierSearchResponse" }
  /suppliers/by-code/{code}:
    get:
      operationId: getSupplierByCode
      tags: [master-data]
      summary: Resolve an authorized supplier code
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Supplier detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupplierDetail" }
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [master-data]
      summary: Get an authorized supplier detail
      parameters:
        - { $ref: "#/components/parameters/IntegerId" }
      responses:
        "200":
          description: Supplier detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupplierDetail" }
  /admin/suppliers:
    post:
      operationId: createSupplier
      tags: [administration, master-data]
      summary: Create a development or provider supplier
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SupplierInput" }
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupplierDetail" }
  /admin/suppliers/{id}:
    put:
      operationId: updateSupplier
      tags: [administration, master-data]
      summary: Update an authorized supplier
      parameters:
        - { $ref: "#/components/parameters/IntegerId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SupplierUpdate" }
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupplierDetail" }
  /admin/suppliers/{supplierId}/aliases:
    post:
      operationId: createSupplierAlias
      tags: [administration, master-data]
      summary: Add a searchable supplier alias
      parameters:
        - { name: supplierId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SupplierAliasInput" }
      responses:
        "201":
          description: Supplier detail after alias creation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupplierDetail" }
  /admin/suppliers/{supplierId}/aliases/{aliasId}:
    delete:
      operationId: deleteSupplierAlias
      tags: [administration, master-data]
      summary: Remove a supplier alias
      parameters:
        - { name: supplierId, in: path, required: true, schema: { type: integer } }
        - { name: aliasId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Supplier detail after alias deletion
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupplierDetail" }
...
components:
  parameters:
...
    IntegerId:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
      required: [error]
...
    MaterialInput:
      type: object
      properties:
        customerId: { type: integer }
        materialCode: { type: string, minLength: 1 }
        materialDescription: { type: string, minLength: 1 }
        shortDescription: { type: ["string", "null"] }
        materialGroup: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        brand: { type: ["string", "null"] }
        baseUom: { type: string, minLength: 1 }
        purchaseUom: { type: ["string", "null"] }
        supplierMaterialCode: { type: ["string", "null"] }
        manufacturerPartNumber: { type: ["string", "null"] }
        plantCode: { type: ["string", "null"] }
        storageLocation: { type: ["string", "null"] }
        batchManaged: { type: boolean }
        expiryManaged: { type: boolean }
        active: { type: boolean }
        sourceSystem: { type: string }
        externalId: { type: ["string", "null"] }
        metadataJson: { type: ["object", "null"], additionalProperties: true }
      required: [customerId, materialCode, materialDescription, baseUom]
    MaterialUpdate:
      allOf:
        - $ref: "#/components/schemas/MaterialInput"
      description: Partial material update; omitted fields are preserved.
...
    MaterialDetail:
      allOf:
        - $ref: "#/components/schemas/MaterialSummary"
        - type: object
          properties:
            aliases: { type: array, items: { $ref: "#/components/schemas/MaterialAlias" } }
            barcodes: { type: array, items: { $ref: "#/components/schemas/MaterialBarcode" } }
            uomConversions: { type: array, items: { $ref: "#/components/schemas/MaterialUomConversion" } }
            supplierMappings: { type: array, items: { $ref: "#/components/schemas/SupplierMaterialMapping" } }
          required: [aliases, barcodes, uomConversions, supplierMappings]
    MaterialAliasInput:
      type: object
      properties:
        alias: { type: string, minLength: 1 }
        aliasType: { type: string, minLength: 1 }
        active: { type: boolean }
      required: [alias, aliasType]
...
    MaterialBarcodeInput:
      type: object
      properties:
        barcode: { type: string, minLength: 1 }
        barcodeType: { type: ["string", "null"] }
        isPrimary: { type: boolean }
        active: { type: boolean }
      required: [barcode]
...
    MaterialUomConversionInput:
      type: object
      properties:
        fromUom: { type: string, minLength: 1 }
        toUom: { type: string, minLength: 1 }
        conversionFactor: { type: string, minLength: 1 }
        sourceSystem: { type: string }
        active: { type: boolean }
      required: [fromUom, toUom, conversionFactor]
...
    MaterialSearchResponse:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/MaterialSummary" } }
        count: { type: integer }
      required: [items, count]
    MaterialRecognitionResponse:
      type: object
      properties:
        strategy: { type: string }
        candidates: { type: array, items: { $ref: "#/components/schemas/MaterialSummary" } }
        recognized:
          oneOf:
            - { $ref: "#/components/schemas/MaterialSummary" }
            - { type: "null" }
      required: [strategy, candidates, recognized]
    MaterialConfirmationInput:
      type: object
      properties:
        materialId: { type: integer }
        clues: { type: object, additionalProperties: true }
      required: [materialId]
...